Added configuration file for server settings.

+++ src/GbxRemote/GbxRemote.h
@@ -8,7 +8,7 @@
 
 struct first_response
 {
-    long size;
+    short size;
 };
 
 struct message_response
